backend/api-gateway/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
beanie==1.27.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lazy-model==0.2.0
motor==3.6.0
odmantic==1.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.10.1
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.32.1
